requirements.txt
Django==4.2.20
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0
django-taggit==6.1.0
django-redis==5.4.0
cryptography==44.0.2
python-dotenv==1.0.1
python-decouple==3.8
requests==2.32.3
tronpy==0.5.0
PyJWT==2.10.1
redis==5.2.1